chore: Rename layout folder
