+++ src/components/PageHeader/PageHeader.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
 import { Box } from '@pancakeswap/uikit'
-import Container from '../layout/Container'
+import Container from '../Layout/Container'
 
 const Outer = styled(Box)<{ background?: string }>`
   background: ${({ theme, background }) => background || theme.colors.gradients.bubblegum};
