+++ src/components/PageSection/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
 import { BoxProps, Box, Flex, FlexProps } from '@pancakeswap/uikit'
-import Container from 'components/layout/Container'
+import Container from 'components/Layout/Container'
 import CurvedDivider from './CurvedDivider'
 
 interface PageSectionProps extends BackgroundColorProps {
