+++ src/views/Collectibles/index.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import styled from 'styled-components'
 import { Heading } from '@pancakeswap/uikit'
 import { useTranslation } from 'contexts/Localization'
-import Page from 'components/layout/Page'
+import Page from 'components/Layout/Page'
 import NftList from './components/NftList'
 
 const StyledHero = styled.div`
