+++ src/views/Home/components/LotteryBanner.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import styled from 'styled-components'
 import { Text, Flex, Button, ArrowForwardIcon, Skeleton } from '@pancakeswap/uikit'
 import { useTranslation } from 'contexts/Localization'
-import Container from 'components/layout/Container'
+import Container from 'components/Layout/Container'
 import { NavLink } from 'react-router-dom'
 import Balance from 'components/Balance'
 import { usePriceCakeBusd } from 'state/hooks'
