+++ src/views/NotFound.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
 import { Button, Heading, Text, LogoIcon } from '@pancakeswap/uikit'
-import Page from 'components/layout/Page'
+import Page from 'components/Layout/Page'
 import { useTranslation } from 'contexts/Localization'
 
 const StyledNotFound = styled.div`
