+++ src/views/Profile/ProfileCreation/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Page from 'components/layout/Page'
+import Page from 'components/Layout/Page'
 import Header from './Header'
 import ProfileCreationProvider from './contexts/ProfileCreationProvider'
 import Steps from './Steps'
