+++ src/views/Teams/index.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { AutoRenewIcon, Flex, Heading } from '@pancakeswap/uikit'
 import orderBy from 'lodash/orderBy'
 import { useTeams } from 'state/hooks'
-import Page from 'components/layout/Page'
+import Page from 'components/Layout/Page'
 import { useTranslation } from 'contexts/Localization'
 import TeamListCard from './components/TeamListCard'
 import TeamHeader from './components/TeamHeader'
