+++ src/views/Voting/CreateProposal/index.tsx
@@ -26,7 +26,7 @@ import { getBscScanAddressUrl, getBscScanBlockNumberUrl } from 'utils/bscscan'
 import truncateWalletAddress from 'utils/truncateWalletAddress'
 import { signMessage } from 'utils/web3React'
 import { useTranslation } from 'contexts/Localization'
-import Container from 'components/layout/Container'
+import Container from 'components/Layout/Container'
 import { DatePicker, TimePicker } from 'components/DatePicker'
 import UnlockButton from 'components/UnlockButton'
 import ReactMarkdown from 'components/ReactMarkdown'
