+++ src/views/Voting/Proposal/index.tsx
@@ -7,7 +7,7 @@ import { ProposalState, VotingStateLoadingStatus } from 'state/types'
 import { useGetProposal, useGetVotingStateLoadingStatus, useGetVotes, useGetProposalLoadingStatus } from 'state/hooks'
 import { fetchProposal, fetchVotes } from 'state/voting'
 import { useTranslation } from 'contexts/Localization'
-import Container from 'components/layout/Container'
+import Container from 'components/Layout/Container'
 import ReactMarkdown from 'components/ReactMarkdown'
 import PageLoader from 'components/PageLoader'
 import { isCoreProposal } from '../helpers'
