+++ src/views/Voting/components/Hero.tsx
@@ -3,7 +3,7 @@ import { Box, Button, Flex, Heading, ProposalIcon } from '@pancakeswap/uikit'
 import styled from 'styled-components'
 import { Link } from 'react-router-dom'
 import { useTranslation } from 'contexts/Localization'
-import Container from 'components/layout/Container'
+import Container from 'components/Layout/Container'
 import DesktopImage from './DesktopImage'
 
 const StyledHero = styled(Box)`
